redpill-backend/requirements.txt
langchain
langchain-nvidia-ai-endpoints
langgraph
qdrant-client
Pillow
python-docx
SpeechRecognition
pydub
streamlit
python-dotenv
langchain-community
sentence-transformers
fastapi
uvicorn[standard]
aiofiles
axios
pytesseract